PR comments addressed.
